requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pillow==11.1.0
reportlab==4.4.0
